# handlers/subscription.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strconv"
	"time"

	"streamshort/models"

	"github.com/gorilla/mux"
	"gorm.io/gorm"
)

type SubscriptionHandler struct {
	db *gorm.DB
}

func NewSubscriptionHandler(db *gorm.DB) *SubscriptionHandler {
	return &SubscriptionHandler{db: db}
}

// GetUserSubscriptions returns all subscriptions for the authenticated user
// GetUserSubscriptions godoc
// @Summary Get User Subscriptions
// @Description Get all subscriptions for the authenticated user.
// @Tags subscription
// @Produce  json
// @Param   page      query    int     false  "Page number"
// @Param   per_page  query    int     false  "Items per page"
// @Success 200 {object} map[string]interface{}
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/subscriptions [get]
func (h *SubscriptionHandler) GetUserSubscriptions(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get query parameters for pagination
	page, _ := strconv.Atoi(r.URL.Query().Get("page"))
	if page <= 0 {
		page = 1
	}
	perPage, _ := strconv.Atoi(r.URL.Query().Get("per_page"))
	if perPage <= 0 || perPage > 100 {
		perPage = 20
	}
	offset := (page - 1) * perPage

	// Get subscriptions with pagination
	var subscriptions []models.Subscription
	var total int64

	// Count total subscriptions
	if err := h.db.Model(&models.Subscription{}).Where("user_id = ?", userID).Count(&total).Error; err != nil {
		http.Error(w, "Failed to count subscriptions", http.StatusInternalServerError)
		return
	}

	// Get paginated subscriptions
	if err := h.db.Where("user_id = ?", userID).
		Preload("User").
		Order("created_at DESC").
		Offset(offset).
		Limit(perPage).
		Find(&subscriptions).Error; err != nil {
		http.Error(w, "Failed to fetch subscriptions", http.StatusInternalServerError)
		return
	}

	// Build response
	response := map[string]interface{}{
		"total":         total,
		"page":          page,
		"per_page":      perPage,
		"total_pages":   (total + int64(perPage) - 1) / int64(perPage),
		"subscriptions": subscriptions,
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}

// GetUserSubscription returns a specific subscription for the authenticated user
// GetUserSubscription godoc
// @Summary Get User Subscription Details
// @Description Get details of a specific subscription.
// @Tags subscription
// @Produce  json
// @Param   id   path     string  true  "Subscription ID"
// @Success 200 {object} models.Subscription
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/subscriptions/{id} [get]
func (h *SubscriptionHandler) GetUserSubscription(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get subscription ID from URL
	vars := mux.Vars(r)
	subscriptionID := vars["id"]

	// Find subscription and verify ownership
	var subscription models.Subscription
	if err := h.db.Where("id = ? AND user_id = ?", subscriptionID, userID).
		Preload("User").
		First(&subscription).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Subscription not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(subscription)
}

// CancelUserSubscription cancels a user's subscription
// CancelUserSubscription godoc
// @Summary Cancel Subscription
// @Description Cancel an active subscription.
// @Tags subscription
// @Produce  json
// @Param   id   path     string  true  "Subscription ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/subscriptions/{id}/cancel [post]
func (h *SubscriptionHandler) CancelUserSubscription(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get subscription ID from URL
	vars := mux.Vars(r)
	subscriptionID := vars["id"]

	// Find subscription and verify ownership
	var subscription models.Subscription
	if err := h.db.Where("id = ? AND user_id = ?", subscriptionID, userID).
		First(&subscription).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Subscription not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Check if subscription can be cancelled
	if subscription.Status != "active" {
		http.Error(w, "Subscription is not active and cannot be cancelled", http.StatusBadRequest)
		return
	}

	// Update status to cancelled
	if err := h.db.Model(&subscription).Update("status", "cancelled").Error; err != nil {
		http.Error(w, "Failed to cancel subscription", http.StatusInternalServerError)
		return
	}

	response := map[string]interface{}{
		"message":         "Subscription cancelled successfully",
		"subscription_id": subscriptionID,
		"status":          "cancelled",
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// RenewUserSubscription renews a user's subscription
// RenewUserSubscription godoc
// @Summary Renew Subscription
// @Description Renew an active or expired subscription.
// @Tags subscription
// @Produce  json
// @Param   id   path     string  true  "Subscription ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Not Found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/subscriptions/{id}/renew [post]
func (h *SubscriptionHandler) RenewUserSubscription(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// Get subscription ID from URL
	vars := mux.Vars(r)
	subscriptionID := vars["id"]

	// Find subscription and verify ownership
	var subscription models.Subscription
	if err := h.db.Where("id = ? AND user_id = ?", subscriptionID, userID).
		First(&subscription).Error; err != nil {
		if err == gorm.ErrRecordNotFound {
			http.Error(w, "Subscription not found", http.StatusNotFound)
			return
		}
		http.Error(w, "Database error", http.StatusInternalServerError)
		return
	}

	// Check if subscription can be renewed
	if subscription.Status != "active" && subscription.Status != "expired" {
		http.Error(w, "Subscription cannot be renewed", http.StatusBadRequest)
		return
	}

	// Calculate new end date (extend from current end date or now)
	var newEndDate time.Time
	if subscription.Status == "active" {
		newEndDate = subscription.EndDate.AddDate(0, 0, 30) // Add 30 days
	} else {
		newEndDate = time.Now().AddDate(0, 0, 30) // Add 30 days from now
	}

	// Update subscription
	updates := map[string]interface{}{
		"status":     "active",
		"end_date":   newEndDate,
		"updated_at": time.Now(),
	}

	if err := h.db.Model(&subscription).Updates(updates).Error; err != nil {
		http.Error(w, "Failed to renew subscription", http.StatusInternalServerError)
		return
	}

	response := map[string]interface{}{
		"message":         "Subscription renewed successfully",
		"subscription_id": subscriptionID,
		"status":          "active",
		"new_end_date":    newEndDate,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// GetSubscriptionPlans returns available subscription plans
// GetSubscriptionPlans godoc
// @Summary Get Subscription Plans
// @Description Get list of available subscription plans.
// @Tags subscription
// @Produce  json
// @Success 200 {array} models.SubscriptionPlan
// @Failure 500 {string} string "Internal Server Error"
// @Router /api/subscriptions/plans [get]
func (h *SubscriptionHandler) GetSubscriptionPlans(w http.ResponseWriter, r *http.Request) {
	var plans []models.SubscriptionPlan

	if err := h.db.Where("is_active = ?", true).Find(&plans).Error; err != nil {
		http.Error(w, "Failed to fetch subscription plans", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(plans)
}

// CheckSubscriptionStatus checks if a user has access to specific content
// CheckSubscriptionStatus godoc
// @Summary Check Subscription Status
// @Description Check if the user has an active subscription.
// @Tags subscription
// @Produce  json
// @Success 200 {object} map[string]interface{}
// @Router /api/subscriptions/check [get]
func (h *SubscriptionHandler) CheckSubscriptionStatus(w http.ResponseWriter, r *http.Request) {
	// Get user ID from context
	userID, ok := r.Context().Value("user_id").(string)
	if !ok {
		http.Error(w, "User ID not found in context", http.StatusInternalServerError)
		return
	}

	// New behavior: user has access if they have any active subscription
	var subscription models.Subscription
	err := h.db.Where("user_id = ? AND status = ?", userID, "active").
		Order("end_date DESC").First(&subscription).Error

	hasAccess := false
	var subscriptionDetails *models.Subscription

	if err == nil && !subscription.IsExpired() {
		hasAccess = true
		subscriptionDetails = &subscription
	} else if err == nil && subscription.IsExpired() {
		h.db.Model(&subscription).Update("status", "expired")
	}

	response := map[string]interface{}{
		"has_access":           hasAccess,
		"target_type":          "global",
		"subscription_details": subscriptionDetails,
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(response)
}